Initial commit of instruction types for 65816 and basic ROM headers
The designed-for-humans assembly mnemonics for 65816 gets in the way of clean
Haskell code (as far as I, not a good Haskell programmer, can do). This gives
us an opportunity to change the types to be less reflective of 65816 syntax and
more reflective of others (to help whom?).

The ROM header file is incomplete (it can't generate a header). I have some
basic fields in there, which don't fully capture all of the various field
values encountered in the headers of real games.

The remaining question is: who will determine the structure of game code and
graphics data in the ROM? Previously, we discussed some notion where the
compiler would hand sections of code and data together for another component of
the compiler to put together. However, to be efficient, the code generator
should know where it is getting the data from, so it should determine where to
fit everything together.